+++ CHANGES.md
@@ -3,6 +3,9 @@ changelog 🧚
 
 ## Unreleased
 
+- Dynamically load parsers with tree-sitter-loader.
+  - Sources go in ~/.cache/dook/sources.
+  - Compiled parsers go in ~/.cache/tree-sitter/lib.
 - Config file format version 2:
   - Add integer "version" field.
   - Move per-language config to live under "languages".